shopnice/requirements.txt
Flask==3.0.2
flask_wtf==1.2.1
passlib==1.7.4
playwright==1.41.2
Requests==2.31.0
pipenv==2023.12.1
WTForms==3.1.2
squareup==33.0.0.20231115
python-dotenv
fastapi
uvicorn
aiofiles
